pyproject.toml
[tool.poetry]
name = "test_project"
version = "0.1.0"
description = "Test task"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
Flask-Script = "^2.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
